#40565 Fix code style
