+++ plugins/org.jkiss.dbeaver.model.sql/src/org/jkiss/dbeaver/model/sql/generator/SQLGeneratorInnerJoin.java
@@ -20,7 +20,6 @@ import org.jkiss.dbeaver.model.DBUtils;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.dbeaver.model.sql.SQLDialect;
 import org.jkiss.dbeaver.model.sql.SQLUtils;
-import org.jkiss.dbeaver.model.sql.format.SQLFormatUtils;
 import org.jkiss.dbeaver.model.struct.DBSEntity;
 
 import java.lang.reflect.InvocationTargetException;
@@ -64,7 +63,7 @@ public class SQLGeneratorInnerJoin extends SQLGenerator<DBSEntity> {
         } catch (Exception e) {
             throw new InvocationTargetException(e);
         }
-        result = this.isFormatSql() ? SQLFormatUtils.formatSQL(objects.getFirst().getDataSource(), sql.toString()) : sql.toString().trim();
+        result = formatIfApplicable(objects.getFirst().getDataSource(), sql);
     }
 
     @Override
